examples: unify doc samples across bindings + fix 3 bugs (#5971)
* examples: unify doc samples across bindings + fix 3 bugs

Align C/C#/Python doc samples to match cpp canonical (same inputs,
same algorithm parameters, same output filenames) where fixture-safe.
C samples lose CLI arg handling in favor of hardcoded filenames,
matching cpp/py convention.

Bugs surfaced during the audit:
- MeshICP.dox.c: refMop was constructed from meshFloating instead
  of meshReference, so ICP was aligning a mesh onto itself.
- MeshLoadSave.dox.cpp: error-check condition was inverted; the
  success path printed the error and exited 1.
- GlobalRegistration.dox.py: p2pl_inaccuracy passed the point
  metric to getMeanSqDistToPlane instead of the plane metric.

Other notable changes:
- C# Collision / CollisionPrecise / CollisionSelf classes renamed
  to *Example so Program.cs reflection dispatcher can invoke them.
- MeshFixDegeneracies.dox.cpp and Triangulation.dox.cpp now save
  their results (were pedagogically incomplete).
- MeshDecimate.dox.cs switched to cpp canonical params; the
  existing S3 fixture was generated with the prior params, so the
  C# MeshDecimate test will fail until the fixture is refreshed.
- MeshModification.dox.py now builds a torus programmatically
  (matching cpp) instead of loading mesh.stl; the corresponding
  parametrize entry in test_python_samples.py drops input_files.


* examples: fix MeshFillHole.dox.c undeclared-function compile error

The original file had `MR_std_vector_MR_EdgeId_Size` / `_At` with
capital S/A inside a dead `#else` branch that was never compiled.
My rewrite inherited that typo and it broke the build on platforms
that no longer silently warn. Use the real lowercase helpers
`_size` / `_at`, matching the convention used across other
c-examples (e.g. MeshStitchHole.dox.c).


* examples: silence -Wformat on macOS in c-examples


* examples: skip MeshDecimate C# output comparison pending fixture refresh


* examples: rename C# MeshDecimate output to decimated_mesh_cs.stl


* examples: hoist FillHoleParams out of the per-edge loop, mention fillHoles


* examples: use literal \xb1 instead of \u00b1 in GlobalRegistration.dox.py


---------

+++ examples/c-examples/Collision.dox.c
@@ -45,8 +45,8 @@ int main( void )
     MR_uint64_t numColB = MR_BitSet_count(
         MR_FaceBitSet_UpcastTo_MR_BitSet(
             MR_std_pair_MR_FaceBitSet_MR_FaceBitSet_second( collidingBitSets ) ) );
-    fprintf( stdout, "%" PRIu64 "\n", numColA ); // print number of colliding faces from mesh A
-    fprintf( stdout, "%" PRIu64 "\n", numColB ); // print number of colliding faces from mesh B
+    fprintf( stdout, "%zu\n", (size_t)numColA ); // print number of colliding faces from mesh A
+    fprintf( stdout, "%zu\n", (size_t)numColB ); // print number of colliding faces from mesh B
 
 
     // fast check if mesh A and mesh B collide
